Revise BDPT renderer for new interface.
