+++ src/renderer/brdf.h
@@ -34,7 +34,6 @@ namespace spica {
     public:
         virtual ~BSDFBase() {}
         virtual Color reflectance() const = 0;
-        virtual Color emittance() const = 0;
         virtual void sample(const Vector3D& in, const Vector3D& normal, double rand1, double rand2, Vector3D* out, double* pdf) const = 0;
         virtual BSDFBase* clone() const = 0;
     };
@@ -47,17 +46,15 @@ namespace spica {
     class SPICA_BRDF_DLL LambertianBRDF : public BSDFBase {
     private:
         Color _reflectance;
-        Color _emittance;
 
     public:
-        static BSDF factory(const Color& reflectance, const Color& emittance = Color(0.0, 0.0, 0.0));
+        static BSDF factory(const Color& reflectance);
         Color reflectance() const override;
-        Color emittance() const override;
         void sample(const Vector3D& in, const Vector3D& normal, double rand1, double rand2, Vector3D* out, double* pdf) const override;
         BSDFBase* clone() const override;
 
     private:
-        explicit LambertianBRDF(const Color& reflectance, const Color& emittance);
+        explicit LambertianBRDF(const Color& reflectance);
     };
 
 
@@ -68,17 +65,15 @@ namespace spica {
     class SPICA_BRDF_DLL SpecularBRDF : public BSDFBase {
     private:
         Color _reflectance;
-        Color _emittance;
 
     public:
-        static BSDF factory(const Color& reflectance, const Color& emittance = Color(0.0, 0.0, 0.0));
+        static BSDF factory(const Color& reflectance);
         Color reflectance() const override;
-        Color emittance() const override;
         void sample(const Vector3D& in, const Vector3D& normal, double rand1, double rand2, Vector3D* out, double* pdf) const override;
         BSDFBase* clone() const override;
 
     private:
-        explicit SpecularBRDF(const Color& reflectance, const Color& emittance);
+        explicit SpecularBRDF(const Color& reflectance);
     };
 
 
@@ -89,18 +84,16 @@ namespace spica {
     class SPICA_BRDF_DLL PhongBRDF : public BSDFBase {
     private:
         Color _reflectance;
-        Color _emittance;
         double _coeff;
 
     public:
-        static BSDF factory(const Color& reflectance, const Color& emittance = Color(0.0, 0.0, 0.0), const double n = 32);
+        static BSDF factory(const Color& reflectance, const double n = 32);
         Color reflectance() const override;
-        Color emittance() const override;
         void sample(const Vector3D& in, const Vector3D& normal, double rand1, double rand2, Vector3D* out, double* pdf) const override;
         BSDFBase* clone() const override;
 
     private:
-        PhongBRDF(const Color& reflectance, const Color& emittance, const double n);
+        PhongBRDF(const Color& reflectance, const double n);
     };
 
     // --------------------------------------------------
@@ -110,17 +103,15 @@ namespace spica {
     class SPICA_BRDF_DLL RefractiveBSDF : public BSDFBase {
     private:
         Color _reflectance;
-        Color _emittance;
 
     public:
-        static BSDF factory(const Color& reflectance, const Color& emittance = Color(0.0, 0.0, 0.0));
+        static BSDF factory(const Color& reflectance);
         Color reflectance() const override;
-        Color emittance() const override;
         void sample(const Vector3D& in, const Vector3D& normal, double rand1, double rand2, Vector3D* out, double* pdf) const override;
         BSDFBase* clone() const override;
 
     private:
-        RefractiveBSDF(const Color& reflectance, const Color& emittance);
+        RefractiveBSDF(const Color& reflectance);
     };
 
 }  // namespace spica
