+++ src/renderer/bsdf.cc
@@ -70,10 +70,6 @@ namespace spica {
         return _ptr->reflectance();
     }
 
-    const Color& BSDF::emittance() const {
-        return _ptr->emittance();
-    }
-
     void BSDF::sample(const Vector3D& in, const Vector3D& normal, double rand1, double rand2, Vector3D* out, double* pdf) const {
         _ptr->sample(in, normal, rand1, rand2, out, pdf);
     }
