+++ src/renderer/scene.h
@@ -32,10 +32,16 @@ namespace spica {
     class SPICA_SCENE_DLL Scene : private Uncopyable {
     private:
         std::vector<Triangle> _triangles;
-        std::vector<Triangle> _lights;
+        std::vector<Color> _emittance;
+
         std::vector<unsigned int> _bsdfIds;
+        std::vector<unsigned int> _lightIds;
+        std::vector<double> _lightPdfs;
+        double _totalLightArea;
+
         std::vector<BSDF> _bsdfs;
         std::shared_ptr<AccelBase> _accel;
+
         Envmap _envmap;
 
     public:
@@ -43,17 +49,26 @@ namespace spica {
         ~Scene();
 
         template <class Ty>
-        void add(const Ty& geom, const BSDF& bsdf, bool isLight = false) {
+        void add(const Ty& geom, const BSDF& bsdf, const Color& emittance = Color(0.0, 0.0, 0.0), bool isLight = false) {
             static_assert(std::is_base_of<IGeometry, Ty>::value, "Type inherits IGeometry can only be added to the scene !!");
 
             std::vector<Triangle> newTriangles = geom.triangulate();
+
+            // If new object is a light, store triangle indices
             if (isLight) {
-                // If geometry is a light, set it to lights
-                _lights.insert(_lights.end(), newTriangles.begin(), newTriangles.end());
+                for (int i = 0; i < newTriangles.size(); i++) {
+                    _lightIds.push_back(_triangles.size() + i);
+                }
             }
 
             // Both normal geoms and light geoms are set to triangles
             _triangles.insert(_triangles.end(), newTriangles.begin(), newTriangles.end());
+
+            // Compute and store emittance
+            for (int i = 0; i < newTriangles.size(); i++) {
+                Color emt = Color(emittance * (newTriangles[i].area() / geom.area()));
+                _emittance.push_back(emt);
+            }
             
             // Update BSDF ids
             const int newBsdfId = static_cast<int>(_bsdfs.size());
@@ -65,9 +80,13 @@ namespace spica {
 
         const Triangle& getTriangle(int id) const;
         const BSDF& getBsdf(int id) const;
+        const Color& getEmittance(int id) const;
+
+        int sampleLight(double rand1) const;
 
         void clear();
         void setAccelerator(AccelType accel = QBVH_ACCEL);
+        void computeLightPdfs();
 
         bool intersect(const Ray& ray, Intersection& isect) const;
 
@@ -75,6 +94,7 @@ namespace spica {
 
         inline const Envmap& envmap() const { return _envmap; }
         inline void setEnvmap(const Envmap& envmap) { _envmap = envmap; }
+        inline double totalLightArea() const { return _totalLightArea; }
     };
 }
