+++ src/utils/path.h
@@ -5,6 +5,13 @@
 #ifndef _SPICA_PATH_H_
 #define _SPICA_PATH_H_
 
+#if _MSC_VER
+#include <direct.h>
+#define mkdir(d) _mkdir(d)
+#else
+#include <sys/stat.h>
+#endif
+
 #include <string>
 
 namespace spica {
@@ -14,6 +21,10 @@ namespace spica {
         inline std::string getExtension(const std::string& filename) {
             const int pos = filename.find_last_of(".");
             return filename.substr(pos);
+        }
+
+        inline bool createDirectory(const std::string& dirname) {
+            return mkdir(dirname.c_str()) == 0;   
         }
 
     }
