+++ src/utils/sampler.h
@@ -31,6 +31,8 @@ namespace spica {
 
         SPICA_SAMPLER_DLL void onTriangle(const Triangle& tri, Vector3D* position, Vector3D* normal);
 
+        SPICA_SAMPLER_DLL void onTriangle(const Triangle& tri, Vector3D* position, Vector3D* normal, double r1, double r2);
+
         SPICA_SAMPLER_DLL void onQuad(const Quad& quad, Vector3D* position, Vector3D* normal);
 
         SPICA_SAMPLER_DLL void onQuad(const Quad& quad, Vector3D* position, Vector3D* normal, double r1, double r2);
